qt6: complete the port - PCM decoder descriptor + Bms encoding/containers
BmsONE now builds on both Qt5 and Qt6 (verified locally with qmake/qmake6).

Audio decoders:
- New audio/PcmFormat.h: portable PCM descriptor (sampleRate/channels/
  sampleSize/sampleType + bytesPerFrame etc.). Qt6 QAudioFormat can only
  represent UInt8/Int16/Int32/Float and dropped sampleSize()/sampleType(),
  so the WAV/OGG decoders use PcmFormat instead; decode logic unchanged.
- Wave.h, DocumentDef.h (WaveSummary), WaveData/WaveStream/
  SoundChannelInternal switched QAudioFormat carrier -> PcmFormat; remaining
  QAudioBuffer::S8U/S8S/S16U stereo typedefs -> StereoFrame<>.

BMS reader:
- QTextStream::setCodec removed in Qt6: decode the file with QTextCodec
  (Core5Compat) into a QString and stream from it (works on both versions).
- QString::indexOf(QRegExp) removed: use QRegularExpression for the
  delimiter; add <QVariant> include; QChar(int) ambiguity and QMultiMap/QMap
  iterator split (Qt6) fixed.

Misc Qt6: qsizetype casts in std::min/max across MasterView/Document/
SoundChannel/WaveStream/SoundChannelInternal.

Verified: Qt5 and Qt6 release builds link; ASan importer harness passes.

+++ src/document/Document.cpp
@@ -562,7 +562,7 @@ void Document::MoveSoundChannel(int indexBefore, int indexAfter)
 {
     if (indexBefore < 0 || indexBefore >= (int)soundChannels.size())
         return;
-    indexAfter = std::max(0, std::min(soundChannels.size()-1, indexAfter));
+    indexAfter = std::max(0, std::min((int)soundChannels.size()-1, indexAfter));
     if (indexBefore == indexAfter)
         return;
     auto updater = [this](int indexOld, int indexNew){
